requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
camelot-py==0.11.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
discord==2.3.2
discord-py-interactions==5.10.0
discord-py-slash-command==4.2.1
discord-typings==0.7.0
discord.py==2.3.2
distro==1.8.0
emoji==2.8.0
et-xmlfile==1.1.0
frozenlist==1.4.0
ghostscript==0.7
idna==3.4
JPype1==1.5.0
multidict==6.0.4
numpy==1.26.4
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.0.0
pycparser==2.21
pypdf==3.16.2
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tabula==1.0.5
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
yarl==1.9.2
